+++ packages/desktop/src/link/Link.tsx
@@ -14,6 +14,7 @@ export interface LinkProps {
   rel?: string
   title?: string
   size?: 's' | 'm' | 'l'
+  bold?: boolean
 }
 
 const LinkSize: { [size in NonNullable<LinkProps['size']>]: number } = {
@@ -66,6 +67,7 @@ export const Link: FC<LinkProps> = (props) => (
         title={props.title}
         size={props.size === undefined ? undefined : props.compact ? LinkHeightCompact[props.size] : LinkSize[props.size]}
         height={props.size === undefined ? undefined : LinkHeight[props.size]}
+        weight={props.bold === undefined ? undefined : props.bold ? 500 : 300}
         children={props.children}
       />
     )}